Change default to be 2 workers
Some of the instability and flakes in CI is believed to be caused by
only have 1 worker. Trying 2 workers as the default, but we might
have to bump up the packet instance to make this work.

If we can increase the packet instance and try out 2 workers for a week,
we should be able to see if this improves stability.