Drop CentOS 7 support completely (#1029)
* Drop CentOS 7 support completely

CentOS 7 was marked deprecated in #1000.  Everyone really needs to be
on RHEL 8 or CentOS 8 at this point.  Drop CentOS 7 support completely
now that a warning has been in place for a bit.

RHEL 8 (or CentOS 8) is required because that's the environment we
would expect customers to use to run the installer.  Any issues with
CentOS 7 are not worth our focus anymore.

* Drop dead code after removing centos 7 support

firewalld is now always in use after dropping CentOS 7.