update tools for running local bmo
Have the run script build a development version of the deployment for
metal3 by looking at the production one and removing the bmo and some
other settings we don't want.

Wait longer for the new pod to run, since we now wait for the init
container to download RHCOS.

Log output.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>