Disable ipa-downloader and pull images from cache when quay is down
We don't need ipa downloader anymore. IPA image is embedded in the
OpenShift release so this is unused. This shaves a few minutes off an
install. It leaves support for using upstream Ironic images with
OpenShift for testing, however.

When mirroring images, this also uses the ironic image we've already
cached instead of pointing to the upstream ironic image on Quay.

For sushy tools and vbmc, we don't want dev environments to get out of
date, but in case of a quay outage we are OK to fail the image pulling
and used cached copies. This obviously won't work in CI where the nodes
are fresh, but at least it makes a dev environment more resilient to
outages.