Remove installation of operator-sdk and dep
This removes operator-sdk installation, we've moved on to 0.17 in BMO.
Is there any reason to keep installing such an old version in
dev-scripts?

I also don't know of any reason to continue to install dep. The less we
try to pull from GitHub the better, it's been so flaky lately.