Miscellaneous dev-scripts clean-ups (#1037)
* Disable ipa-downloader and pull images from cache when quay is down

We don't need ipa downloader anymore. IPA image is embedded in the
OpenShift release so this is unused. This shaves a few minutes off an
install. It leaves support for using upstream Ironic images with
OpenShift for testing, however.

When mirroring images, this also uses the ironic image we've already
cached instead of pointing to the upstream ironic image on Quay.

For sushy tools and vbmc, we don't want dev environments to get out of
date, but in case of a quay outage we are OK to fail the image pulling
and used cached copies. This obviously won't work in CI where the nodes
are fresh, but at least it makes a dev environment more resilient to
outages.

* Remove installation of operator-sdk and dep

This removes operator-sdk installation, we've moved on to 0.17 in BMO.
Is there any reason to keep installing such an old version in
dev-scripts?

I also don't know of any reason to continue to install dep. The less we
try to pull from GitHub the better, it's been so flaky lately.

* Only wait for ipa when needed