update tools for running local bmo (#1034)
* update tools for running local bmo

Have the run script build a development version of the deployment for
metal3 by looking at the production one and removing the bmo and some
other settings we don't want.

Wait longer for the new pod to run, since we now wait for the init
container to download RHCOS.

Log output.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* make it possible to override the logging destination directory

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* add stop script for dev metal3 deployment

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* add script to bring mao back under cvo control

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>