ipv6: Enable dual-stack feature set
Dual-stack IPv6 is not yet enabled by default in OpenShift since it
doesn't fully work yet.  We have a feature set that allows you to turn
it on for testing purposes.  When dev-scripts is configured to bring
up a dual-stack test environment, this change adds a manifest to
enable dual-stack in the cluster.

Now that a lot of prerequisite work has started merging, we're ready
to push the dual-stack CI jobs further and this change will ensure
that dual-stack is enabled in those jobs.