be more careful with cleaning up ssh known_hosts (#1049)
The sed calls for removing entries from the ssh known_hosts config
assume that there are v4 and v6 addresses to be removed. If one isn't
found the pattern causes all of the entries from ~/.ssh/known_hosts to
be removed. Protect those sed calls with checks for non-empty
variables.

Also fix a typo in the reference to EXT_SUB_V6 in the invocation for
IPv6.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>