Use IPv4 VIPs for dual-stack envs (#1062)
When running a dual-stack OpenShift cluster, the cluster uses IPv4 as
the primary address family.  We should use IPv4 for the VIPs if we're
only going to use a single address family.

This came up because baremetal-runtimecfg uses the address family of
the API VIP to determine what address to look for on the Node to give
to kubelet to use as the node IP.  Dan Winship suggested that IPv4
would be a better default since that's the primary address family for
the cluster.

This is fine for a short term change, but what we really need to do is
fix this env so that we have both IPv4 and IPv6 VIPs at the same time,
but that's going to take more work and can't be solved only in
dev-scripts alone.