Handle empty assets/generated dir (#1078)
* Handle empty assets/generated dir

It's possible for this to be empty, e.g if we fail the resolution
of clock.redhat.com no custom DNS manifests get created

This appears to currently be the case in CI, so we fail like:

cp: cannot stat 'assets/generated/*.yaml': No such file or directory

* Don't discard output from host clock.redhat.com

It seems this is failing in CI, but it's not clear why -
it should be reasonable to just output the result, and
hopefully this will also expose the reason we're not setting
NTP_SERVERS in CI.

* Fall back NTP server to pool.ntp.org

Since clock.redhat.com is no longer externally resolving we can
fall back to pool.ntp.org

I've not replaced the clock.redhat.com reference as IIRC traffic
to external NTP servers may not work in some test environments,
so in those cases we probably want to stick with the current server.

* Remove the wait-for install-complete workaround

This timeout got increased in the installer for baremetal,
and this currently means we wait doing nothing in CI for
an additional 105mins even when a failure e.g bootstrap timeout
already occured.