Allow testing with custom base images
One problem we have at the moment is testing with custom built
images using the same repositories and configuration from Openshift
upstream and downstream customization.
This patch introduces a bsae-image container that can be used to build
local images using for example custom repositories.
To be able to do that, we introduce 1 new variable that can be set
in the config file: CUSTOM_REPO_FILE
The CUSTOM_REPO_FILE is the name of a file that contains the custom
repositories that we want to use to install packages from, e.g.
export CUSTOM_REPO_FILE=ocp46.repo
The file needs to be present in the base-image dir.