add a script to get the image settings for the active MAO (#1080)
Sometimes when working with the machine-api-operator it is not
necessary to provide a custom set of images, and the images file in
the machine-api-operator repository may not match the cluster where
development is happening. This new script extracts the image settings
from the running machine-api-operator so they can be used with a
custom build.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>