Fix COPY instruction for custom repositories file (#1083)
The paths of files and directories in the COPY instructions inside
the Dockerfile are interpreted as relative to the source of the
context of the build.
Therefore, the custom repository file must be created inside the
base-iamge directory or the build will fail.
Also changing ' to " in sed command to correctl expand the
${BASE_IMAGE_DIR} variable.