Create all workers with a single ansible run (#1085)
Currently we invoke the ansible roles twice, which results
in missing DHCP reservations in the ostestbm network configuration,
because that network already exists when the second run takes place.

That means on scale out any workers defined via NUM_EXTRA_WORKERS will
fail to join the cluster, since they register as "localhost", but the
machine hostname is "localhost.localdomain", which the machine-approver
rejects.