fetch a ci token as needed for the pull secret
The user-specific pull secret granted by logging in to
api.ci.openshift.org expires. Rather than forcing users to manually
get a new one, use the login command to get it ourselves and combine
it with the more static parts of the pull secret.

This change deprecates the PULL_SECRET variable in favor of
PERSONAL_PULL_SECRET and CI_TOKEN variables.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>