fetch a ci token as needed for the pull secret (#1093)
* fetch a ci token as needed for the pull secret

The user-specific pull secret granted by logging in to
api.ci.openshift.org expires. Rather than forcing users to manually
get a new one, use the login command to get it ourselves and combine
it with the more static parts of the pull secret.

This change deprecates the PULL_SECRET variable in favor of
PERSONAL_PULL_SECRET and CI_TOKEN variables.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* use a file for the pull secret

Ask the user for their personal pull secret in a file.

Combine all of the pull secrets, including for the local registry,
into one file and use that file always when passing credentials to
`oc` commands.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* simplify merging of pull secrets

Assume all of the files with pull secrets have the same format and
ignore the fact that we only care about the secret for one host from
the CI registry login step.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* do not show the PULL_SECRET variable when writing to a file

Avoid logging the PULL_SECRET when we write it to the file.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* always combine the PULL_SECRET with local registry credentials

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* remove unused temporary file

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>