have stop-mao.sh save the image settings (#1100)
When running the MAO locally, the developer typically needs a JSON
file with the images the MAO should use. It is possible to build a
custom file, but it's often easier to just use the one from the
cluster because we know those images match the rest of the components
in the cluster.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>