remove support for PULL_SECRET variable (#1109)
* remove support for PULL_SECRET variable

Previous changes moved to using a CI_TOKEN for access to the openshift
CI image registry and a file for the other parts of the pull
secret. The PULL_SECRET variable was retained to allow the CI job to
be updated. That has been done, so now we can remove support for
PULL_SECRET.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* improve error messages when using old configuration mode

Add some extra help text in the error messages produced when the pull
secret and token variables are not set up in the new way.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>