reorganize existing pre-run validation
Some of the validation steps only make sense when we're trying to
build a cluster. For example, there is no reason to require a valid
token or pull-secret when deleting a cluster. So, move the validation
from common.sh into its own file, define functions to invoke the
validations we need in different phases, then add those function calls
to the relevant files.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>