Line-buffer log output
Because there is a pipe (from awk to tee) in our log output pipeline,
output is now buffered, which is bad for watching in real time. awk
doesn't provide an option to line-buffer the output even though it is
not going to a terminal, so explicitly call fflush() after each line.

Fixes #1121