Only add timestamps in logfile, not on console
The make realclean script includes calls to podman that prompt on the
console. Since the prompts don't end with a newline, they are not
printed immediately as stderr is redirected along with stdout.

Move the awk invocation that adds timestamps to the log file path only
(after the tee), so that prompts are displayed on the console
immediately.