Use appropriate ignition version for chronyd config (#1118)
* Use appropriate ignition version for chronyd config

The version from 4.6 onwards switched to ignition 3.1.0 so
we should update the version, the schema for the files
changed slightly so we have to remove the filesystem for
3.x but otherwise the existing config should work.

Also note the default behavior for overwrite changed from
true (2.x) to false (3.x) so we now need to specify this
explicitly to ensure the same behavior in both cases.

* Update file_example.ign to 3.1.0

This is just an example so it's probably reasonable
to provide one that only works with the latest releases,
users are free to copy/modify as needed when testing
with older versions.