Line-buffer log output (#1123)
* Line-buffer log output

Because there is a pipe (from awk to tee) in our log output pipeline,
output is now buffered, which is bad for watching in real time. awk
doesn't provide an option to line-buffer the output even though it is
not going to a terminal, so explicitly call fflush() after each line.

Fixes #1121

* Only add timestamps in logfile, not on console

The make realclean script includes calls to podman that prompt on the
console. Since the prompts don't end with a newline, they are not
printed immediately as stderr is redirected along with stdout.

Move the awk invocation that adds timestamps to the log file path only
(after the tee), so that prompts are displayed on the console
immediately.