validation improvements (#1122)
* reorganize existing pre-run validation

Some of the validation steps only make sense when we're trying to
build a cluster. For example, there is no reason to require a valid
token or pull-secret when deleting a cluster. So, move the validation
from common.sh into its own file, define functions to invoke the
validations we need in different phases, then add those function calls
to the relevant files.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* add earlier validation of CI registry token

When not running in CI, the CI_TOKEN value should be checked early in the
process of building to avoid waiting for a long time before discovering
that it is invalid.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>