Cleanup /var/lib/libvirt/openshift-images
For some reason the virsh --remove-all-storage isn't actually removing
the storage, this may be because we moved to a non-default pool?

This can quickly use up space as each leaked directory is about 2.2G
so lets clean it up until we can find a more robust way to handle
this in the installer.