Fetch oc tools from different URL (#1164)
The current URL we are using for `oc` was removed, so we need to get the
tools from another location. This also removes the dependency on
calculating the version based on the pullspec, meaning this code works
for CI builds too.

We only use these tools to bootstrap getting the ones from the release
payload, so the actual version isn't super important as we just
overwrite it anyway a little bit later on.