update local-bmo.sh to work with cluster-baremetal-operator (#1174)
* add a script to stop the cluster-baremetal-operator

When we are developing the baremetal-operator we may want to stop the
CBO so we can change the resources it manages.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* local-bmo.sh: stop cluster-baremetal-operator instead of machine-api-operator

The machine-api-operator no longer interferes with our ability to run
the baremetal-operator outside of the cluster.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* local-bmo.sh: use external provisioning IP for URLs

The variables in the Deployment for metal3 are not correct when
running the baremetal-operator outside of the cluster, so adjust
them.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* when scaling down metal3, do not wait for image-cache pods to be deleted

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* have cvo ignore host crd as well as metal3 deployment

We probably want to apply our local CRD, so tell the CVO not to
overwrite it.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* always pull images for development mode

When we create the metal3-development deployment, configure it to
always pull images. This may take a few seconds more when the pod
starts up, but it ensures that we have good copies of the images being
used, in case they're local or custom.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>