Add support for infinite DHCP leases
The proper way to do this would be to set the lease length in the
libvirt network config, but unfortunately the version of libvirt we
have in CentOS 8 right now doesn't support that. This adds a script
that manually modifies the libvirt dnsmasq configuration to make the
leases infinite. It is started in the background, and killed when
make clean is run. It can be enabled by setting INFINITE_LEASES=1.
Currently, due to a bug with libvirt's handling of infinite leases,
this will cause make to fail. The cluster will deploy fine, but
some of the post-deploy steps don't work. This is "normal" and
should be fixed in future versions of libvirt.

Background: This is to support testing and dev of a feature that
converts infinite DHCP leases to static configuration so a node no
longer has a dependency on the DHCP server.