Fix clouds.yaml IP when provisioning network disabled
Currently this picks the .8 IP on the controlplane network, but
instead we need to reference the host/pod IP where the metal3
pod is running in the case where there is no provisioning network.

This should also work fine when we do have a provisioning network
so just update to always check the IP from the metal3 pod.

Note that if the metal3 pod gets rescheduled it will be necessary
to run generate_clouds_yaml.sh to update the clouds.yaml