Fix error building images without extra pkgs (#1201)
The extra pkg change broke image builds that don't use it because
it was unconditionally calling realpath on the variable. If it wasn't
set then this resulted in an invalid command. To fix this we just
need to move the realpath call inside the -n check so we only do it
if the variable is populated.