Script to install assisted-installer's operator
An operator has been created and released for the assisted-installer,
which simplifies the deployment process for it and allows us to easliy
run it as part of a dev-script deployment without other node
reqquirements.

This script uses the assisted-service operator to create an AI
deployment. A few things are required:

- Local Storage operator (because AI requires 2 PVs)
- Hive operator (so that the ClusterDeployment CRDs are installed)

Signed-off-by: Flavio Percoco <flavio@redhat.com>