NO-ISSUE Add VM_EXTRADISKS_LIST and VM_EXTRADISKS_SIZE
This PR adds the ability to configure VM_EXTRADISKS_LIST and
VM_EXTRADISKS_SIZE that are required by the Assisted Service
to deploy the Local Storage Operator.

Configuration of the extra disks is automatically picked by the script
so that `storageClassDevices` in the `LocalVolume` manifest is filled
correctly with all the additional drives and does not require a manual
modification.

Co-authored-by: Lisa Rashidi-Ranjbar <lranjbar@redhat.com>