NO-ISSUE Store artifacts in saved-assets/assisted-installer-manifests (#1247)
This PR makes the `assisted_deployment.sh` script save all the manifests
into `saved-assets/` so that operators have an ability to inspect them
after (un)successful installation.

The previous behavior was to generate the manifest on fly and pass it
directly to `oc apply`. From now on the manifest will be saved on a
local filesystem and only afterwards applied from the file that has been
created.

This PR creates as well a patched copy of `extra_host_manifests.yaml`
given that the namespace of the created resource must match the
namespace of all the other resources we are creating.