Improve script to install assisted-service (#1269)
This PR adds additional ways to use `assisted_deployment.sh` so that it
can not only perform a complete installation of the Assisted Service but
can also be used to install only its prerequisites, i.e. LSO and Hive.

This is a useful workflow for development where operator-sdk is used to
install the operator itself, but we don't want to handle the base
requirements manually.