Update base-image logic
Introduce a variable to explicitly trigger the build of a custom base
image for the ironic images.
Do not make that depends from the CUSTOM_REPO_FILE variable anymore.
The CUSTOM_REPO_FILE is not mandatory anymore to build a custom base
image.
Also a custom base image is not required anymore to build local images as
the openshift base images are pullable from the openshift registry.