Do not remove upstream DNS configuraton in cleanup
This PR changes the logic of host cleanup so that the upstream DNS
configuration is not removed. This comes from the fact that by default
dev-scripts configure NetworkManager to use dnsmasq with DHCP-provided
DNS servers and removes the existing /etc/resolv.conf configuration.

This causes problems in environments configured without DHCP with DNS
servers configured manually without use of the NetworkManager. With this
PR user can use `ADDN_DNS` variable to provide their own DNS servers
that will be passed to the dnsmasq. At the same time when removing
dev-scripts from the system, the backup of the original /etc/resolv.conf
will be restored so that the initial system configuration is reverted.