Provide option to enable NATv6
With newer versions of libvirt available in the advanced-virtualization
repository, it's possible to use NATv6. In Packet, each host is given 1
public IPv6 address, so with NATv6 it's possible to make a connected
cluster.

For IPv6 alone, this doesn't help much as Quay still isn't accessible.
However, in dualstack cases, this makes a lot of sense and lets us run
more tests that verify IPv4 and IPv6 connectivity.

This change adds a new configuration option `ENABLE_NATV6` that when set
to true will configure libvirt to use this. This is an opt-in feature as
some lab machines may have IPv6 addresses but not actually have outbound
internet access.