Update MetalLB README file and Makefile rules to reflect the latest configuration changes (#1311)
* Remove MetalLB namespace yaml as it is not used any more

As we are now deploying MetalLB through the operator, no need
to create the namespace from the yaml.

* Update MetalLB configuration Makefile rule

As we no longer configure MetalLB with protocol (the E2E tests are
doing the configuration), there is no need to have these rules.

* Update MetalLB REDAME file to reflect the latest configuration changes

Updating the configuration explanation and adding a section about
E2E tests.
Removing the protocol configuration part as we are not configuring
the protocol through dev-scripts any more.

Removing the service sample yaml as the E2E tests are starting a
service while running.