Add support for mirroring OLM operators
In disconnected deployments (like ipv6) we need to mirror any OLM
operators that we want to install because the OLM images aren't
otherwise available. This patch adds a MIRROR_OLM configuration
variable that will create a pruned index containing only the
relevant operators in the local registry.

The mirroring process generates a couple of manifests needed to use
the mirrored index, and these files are added to the deployment
manifests directory so it will be usable immediately.