Add support for mirroring OLM operators (#1332)
* Add support for mirroring OLM operators

In disconnected deployments (like ipv6) we need to mirror any OLM
operators that we want to install because the OLM images aren't
otherwise available. This patch adds a MIRROR_OLM configuration
variable that will create a pruned index containing only the
relevant operators in the local registry.

The mirroring process generates a couple of manifests needed to use
the mirrored index, and these files are added to the deployment
manifests directory so it will be usable immediately.

* Install opm client when mirroring OLM images

This is required in order to prune the image list so we don't pull
down all of the OLM operator images. It's a fairly simple install,
so let's automate it so there's no additional steps needed to use
the feature.

* Document MIRROR_OLM config value