Add ENABLE_VIRTUAL_MEDIA_VIA_EXTERNAL_NETWORK environment variable (#1357)
This new ENABLE_VIRTUAL_MEDIA_VIA_EXTERNAL_NETWORK will be used to pass
virtualMediaViaExternalNetwork as true in provisioning configs.

virtualMediaViaExternalNetwork, when it is set to true, allows for worker
to boot via Virtual Media and contact metal3 over the External Network.

With this PR, we will be able to test this functionality in our CI
jobs to assure that it is working as expected or not. Moreover,
users also can use that environment variable if they want to boot workers
via Virtual Media and contact metal3 over the External Network.