Silence git checkout of master
On metal3-dev-env where the default branch has been renamed to main,
this command causes an ugly error message in the logs that has
resulted in a lot of wasted debug time. Since we expect that this
command is going to fail some of the time, let's suppress that
output. If there is a legitimate failure on this line, the second
command will also fail so we'll know where to look.