network: Add support to dualstack
The dualstack feature requires openshift 4.12 or higher and set variable IP_STACK
to v4v6.

Example:
   $ devscript> export OPENSHIFT_RELEASE_STREAM=4.12
   $ devscript> export IP_STACK=v4v6
   $ devscript> make all

To verify, check if both VIPs are configured:
   $ devscript> oc get nodes
   $ devscript> oc debug node/master-0.ostest.test.metalkube.org
   $ chroot /host
   # cat /etc/keepalived/keepalived.conf

Please note, as mentioned above, dualstack feature is only available
in openshift 4.12 or higher versions. Based on that, testing this
feature in previous versions will make the installer fail (as expected).
For setting back the environment for IPv4 only, set IP_STACK back to v4.

Example:
   $ devscript> export OPENSHIFT_RELEASE_STREAM=4.11
   $ devscript> export IP_STACK=v4

Depends-on: https://github.com/openshift-metal3/dev-scripts/pull/1462
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>