Agent: change input/output directories (#1406)
* Agent: Put ZTP manifests in cluster-manifests

We want to save the manifests directory for adding extra manifests; ZTP
manifests will go in the cluster-manifests directory. The installer
currently reads them from manifests, so temporarily add a symlink.

* Agent: Expect ISO output in working directory

In order to allow us to remove the 'output' directory, expect the
agent.iso to be directly in the installer's work directory.

Since this hasn't been implemented yet, create a symlink to the ISO in
the output directory. In future the installer will write the ISO over
this symlink, and at that point we can stop creating it.