Add DNS custom host support
To support custom networkConfigs that include static IPs (which
require reverse DNS records for each node so a hostname can be
assigned), this adds the ability to specify additional DNS hosts
that will be added to the regular list of hosts that dev-scripts
creates in DNS.

This config is only added if NETWORK_CONFIG_FOLDER is set and the
file 'hosts.yaml' exists in that location. The file contents should
be a list of IPs and hostnames, such as:

- ip: 192.168.111.110
  hostnames:
    - "master-0"
- ip: 192.168.111.111
  hostnames:
    - "master-1"
etc.