Ensure IP_STACK in agent scenario is always lowercase
I can't explain why it makes any difference, but using the suggested
scenario values of *_IPV4 doesn't work - IPv6 gets configured somewhere
and network connectivity becomes patchy as it round-robins between both
addresses. The process completes slowly or not at all.

In principle it shouldn't matter, because we are setting nocasematch.
Perhaps we are calling another shell process somewhere (metal3-dev-env?)
and that option isn't having an effect there, although I can't find it.
In any event, sticking to lowercase works and uppercase does not.

Since the nocasematch was added presumably for this case, remove it as
well to avoid other stuff starting to rely on it.