Don't append to existing dnsmasq config (#1420)
Previously, if you ran two builds in a row without cleaning in between,
this would append a second copy of the dnsmasq config to the end of the
file. The presence of duplicate line in the config caused an error that
prevented NetworkManager from doing DNS lookups. This meant that even
cleaning was broken, so there was no obvious way to recover.

Overwrite the file instead of appending to it so that this doesn't
happen.