For agent mirror testing force disconnect from masters
This simulates a disconnect from the masters when using the
agent with mirroring. It sets a bogus /etc/hosts entry so that the local
mirror mut be used.  This is currently needed since we must have a valid
connection to create the mirror and create the iso so we can't break the
connection until after that's done.