Add Alma and Rocky Linux as supported host OSes (#1348)
These two distributions are essentially what CentOS was pre-Stream,
so they should generally be compatible as well. They are documented
as "best effort" support though since I don't think we'll be running
them in CI.

For Alma, Ansible does not correctly detect the python interpreter
so we need to explicitly override it on calls to ansible-playbook.
We can't do it in a config file because it's explicitly set in the
metal3-dev-env inventory, so we need to use a method that will take
precedence over that. This workaround can be removed once Ansible
has proper support for Python detection on Alma.