Add interface for custom networkConfig (#1418)
* Add interface for custom networkConfig

This provides a way to specify a custom networkConfig for each node
in a deployment. The user sets NETWORK_CONFIG_FOLDER to a path that
contains YAML files named for each node. The contents of these files
are included in the install-config definition for the appropriate
node.

* Add DNS custom host support

To support custom networkConfigs that include static IPs (which
require reverse DNS records for each node so a hostname can be
assigned), this adds the ability to specify additional DNS hosts
that will be added to the regular list of hosts that dev-scripts
creates in DNS.

This config is only added if NETWORK_CONFIG_FOLDER is set and the
file 'hosts.yaml' exists in that location. The file contents should
be a list of IPs and hostnames, such as:

- ip: 192.168.111.110
  hostnames:
    - "master-0"
- ip: 192.168.111.111
  hostnames:
    - "master-1"
etc.

* Add support for bonding the primary interface

In order to test deployment on bonds we need a way to have two
interfaces on the baremetal network. This adds a new environment
variable, BOND_PRIMARY_INTERFACE, that adds two NICs on the baremetal
network to each node.

* Document network config and bond parameters

Also makes NETWORK_CONFIG_FOLDER able to handle relative paths.