Add static and bond network configs (#1457)
These are some example configs that can be used with the
NETWORK_CONFIG_FOLDER option added in #1418. In addition to making
it easy for new users to test the networkConfig feature, this will
also make it easier to run such configurations in CI.

To use, just set the variable to the appropriate path. For example:

export NETWORK_CONFIG_FOLDER=network-configs/static