AGENNT-316: Add support for oc mirror (#1435)
Add a new variable - OC_MIRROR and support for using 'mirror-registry'
to create a local mirror registry and 'oc mirror' to do the actual
mirroring. This also requires MIRROR_IMAGES to be set. If OC_MIRROR
is not set then the mirroring will be done using 'oc adm release mirror'.

Note that OC_MIRROR is currently only support for the Agent
implementation.

The registry auth must also be added to the user's .docker/config.json
file.