Check for existing cluster before deployment
One thing I've noticed is that a lot of new dev-scripts users don't
realize they need to run make clean between deployments. This adds
a check for the ocp/<cluster name> directory before deployment and
stops if it already exists. That directory is created pretty early
in the process and removed by make clean so it should be a good
indicator of whether a deployment has already been run. It's also
specific to a given cluster so it shouldn't affect environments that
are running multiple clusters of different names.