Fix Assisted Service cleanup script
This PR fixes the script responsible for cleaning up the Assisted
Service deployment. It previously relied on the variable
`NODES_PLATFORM` that has a value of `libvirt` and is unrelated
to how the Assisted Service is deployed. With this change the
Make target `assisted_deployment_cleanup` works correctly again.

This PR changes the generic `clean` target as Assisted cleanup is
redundant at this point as underlying OCP cluster has already been
deleted.