AGENT-392: Add support for install-config.yaml/agent-config.yaml (#1471)
* For agent, if RELEASE_IMAGE is a digest don't set override

if the RELEASE_IMAGE is a digest it should be used as is
since registries.conf will have mirror-by-digest-only-true.

* AGENT-392: Add support for install-config.yaml/agent-config.yaml

Add support for install-config.yaml and agent-config.yaml for
agent deployments and use it by default. ZTP manifests can
still be used by setting the new 'AGENT_USE_ZTP_MANIFESTS' to true.
This uses ansible and jinja to create the yaml files files based
on https://github.com/openshift-metal3/dev-scripts/pull/1458.

* fix for ca-bundle and added mce support

* fix for oc-mirror and mce support

* minor changes needed after VIP fix

* added function for install-config

* support multiple vips

* for dualstack added ipv6 route

* minor changes for VIP settings

* fixes for SNO