AGENT-435: Create ZTP manifest files using Ansible
Move the creation of the ZTP manifest files to Ansible, similar to
what is currently done for registries.conf and install/agent-config.yaml.
A common Ansible role has been added to share vars and templates between
the ZTP manifests and the install/agent-config.yaml.

A follow-on PR will add a role for mce and move files under there.
In addition a follow-on PR will move the setting of variables
currently done in 04_agent_configure.sh to Ansible.