AGENT-429: Dynamically generate mirror configuration for agent (#1478)
The mirror registries configuration for the agent is currently
hardcoded. This is incorrect when a particular release image
digest is defined, for example here is the ICSP for recent 4.12 image.
imageContentSources:
- mirrors:
  - virthost.ostest.test.metalkube.org:5000/localimages/local-release-image
  source: registry.ci.openshift.org/ocp/4.12-2022-11-11-131611
- mirrors:
  - virthost.ostest.test.metalkube.org:5000/localimages/local-release-image
  source: registry.ci.openshift.org/ocp/release

The mirror registries should be extracted from the log files when the
'oc adm release mirror' or 'oc mirror' commands are run. This will be
correct for all images and is easier to maintain.

For the ZTP manifests an extra conversion is needed to put it in the
expected format for registries.conf.