do not stomp on existing yq (#1493)
This change avoids installing the python version of yq if there is
already another version installed. This allows a user to install the
golang version from https://github.com/mikefarah/yq, for example, when
that version is required by other apps or scripts running on the same
host as dev-scripts.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>