Add configurable step to oc-mirror to generate tarfile
Add the ability to use oc-mirror to create a mirror registry by
generating a tarfile first and then publishing the tarfile to the
registry. The can be used to test a common use case for disconnected
installs.

In addition, a function to create a tarfile with only images needed
to bootstrap a host is included. This will be used for testing future
functionality to remove the external registry.