WIP: Add support for doing mirroring from host
Adds a new env variable - AGENT_REGISTRY_ON_NODE. When set to
`true` and MIRROR_IMAGES is also `true`, instead of creating
a registry and mirroring images on the dev-scripts machine, the
images will be mirrored to a tarfile, the tarfile transferred to
a host, and the host will run a registry to serve the images.

This is used as a POC for AGENT-262. It currently only works
for the bootstrap images and requires a new agent installer
service to start the registry and push the image tarfile.