Update dev-scripts CI image
We temporarily need to support both RHEL 7 and 8
as we can't change the ci-operator image build to switch
to rhel a new image in lockstep. Once the switch over is done
we can remove the RHEL7 support.

This should allow use to eventually remove the NSS_WRAPPER_PASSWD
hack in baremetalds-packet-setup-commands.sh