metallb: Override external frr image
This commit adds the --external_frr_image flag
into metallb `inv e2etest` command, thus
overrides the external containers` frr image.
This is required as we're facing e2e test failures
with  FRR's BFD + echo test case, and to prevent
it we want to align the FRR version between
the speakers and the external container.

Signed-off-by: liornoy <lnoy@redhat.com>